can't do imports in d.ts
