+++ components/amorphic-contracts/index.d.ts
@@ -1,6 +1,3 @@
-import { Enums } from './logger-enums';
-import { Interfaces } from './logger-interfaces';
-
 /**
  * application configuration store object
  *
@@ -8,7 +5,7 @@ import { Interfaces } from './logger-interfaces';
  * set => Sets a key value pair for this config object
  * loadFile => loads a file into config object
  */
-export interface Config {
+ export interface Config {
     get(key: string): any
     set(key: string, value: any): any
     loadFile(fileKey: string, file: string): any
@@ -137,4 +134,187 @@ export interface HavenUILogger extends Interfaces.BunyanBaseInterface, HavenLogg
      * @returns {HavenUILogger}
      */
     childLogger?(rootValues?: Interfaces.ChildLogRootValues, dataValues?: Interfaces.ChildLogDataValues): HavenUILogger;
+}
+
+/**
+ * Logger enums
+ */
+ export namespace Enums {
+    export enum Category {
+        security = 'security',
+        availability = 'availability',
+        request = 'request',
+        milestone = 'milestone'
+    }
+
+    export enum PiiLevelEnum {
+        's1' = 's1',
+        's2' = 's2',
+        'nonPII' = 'non-pii'
+    }
+
+    export const enum GlobalHeader {
+        HT_GLOBAL_ID = 'HT-Global-Id',
+        HT_GLOBAL_SOURCE = 'HT-Global-Source'
+    }
+    
+    export type GlobalHeaderType = {
+        [key in GlobalHeader]?: string;
+    }
+
+    export const enum ServerType {
+        express = 'express',
+        restify = 'restify'
+    }
+
+    export enum LogLevel {
+        info = 'info',
+        error = 'error',
+        debug = 'debug',
+        warn = 'warn'
+    }
+    
+    export const LevelFromName: {
+        trace: 10,
+        debug: 20,
+        info: 30,
+        warn: 40,
+        error: 50,
+        fatal: 60
+    };
+}
+
+/**
+ * Logger types and interfaces
+ */
+export namespace Interfaces {
+    export interface BunyanBaseInterface {
+        info(): boolean;
+        error(): boolean;
+        debug(): boolean;
+        warn(): boolean;
+        info(object: ExpandedLog, ...params: any[]): void;
+        info(message: string, functionName: string, piiLevel?: Enums.PiiLevelEnum): void;
+        error(object: ExpandedLog, ...params: any[]): void;
+        debug(object: ExpandedLog, ...params: any[]): void;
+        warn(object: ExpandedLog, ...params: any[]): void;
+    }
+    
+    export type Data = {
+        [key: string]: any;
+    }
+    
+    export type strNumKeyValuePair = {
+        [key: string]: number | string;
+    }
+    
+    export interface ErrorData {
+        isHumanRelated: boolean;
+        code?: string;
+        error?: any;
+        message?: string;
+        data?: Data & { fromError?: never };
+    }
+    
+    export interface Context {
+        globalId?: string,
+        globalSource?: string,
+        referenceId?: string,
+        referenceType?: string,
+        sessionId? : string,
+        piiLevel?:  Enums.PiiLevelEnum,
+        data?: Data
+    }
+    
+    export interface ErrorLog {
+        isHumanRelated?: boolean,
+        code?: string,
+        name?: string,
+        message?: string,
+        stack?: string,
+        data?: Data
+    }
+    
+    interface GeoIP  {
+        latitude?: number,
+        longitude?: number,
+        cityName?: string,
+        stateName?: string,
+        postalCode?: string,
+        timeZone?: string,
+        countryName?: string,
+        countryCode?: string,
+        continentCode?: string
+    }
+    
+    export interface Request {
+        url?: string,
+        method?: string,
+        clientIpAddress?: string,
+        data?: Data,
+        remoteUser?: string,
+        remoteAddress?: string,
+        line?: string,
+        length?: number,
+        time?: number,
+        headers?: string,
+        status?: number,
+        bytesSent?: number,
+        bodyBytesSent?: number,
+        httpReferer?: string,
+        httpUserAgent?: string,
+        httpXForwardedFor?: string,
+        upstreamAddress?: string,
+        upstreamStatus?: number,
+        upstreamResponseTime?: number,
+        upstreamConnectTime?: number,
+        upstreamHeaderTime?: number,
+        upstreamBytesReceived?: number,
+        upstreamBytesSent?: number,
+        geoip?: GeoIP
+    }
+    
+    export interface Response {
+        status?: number,
+        data?: Data
+    }
+    
+    export interface Log {
+        module?: string,
+        function?: string,
+        category?: 'security' | 'availability' | 'request' | 'milestone',
+        message?: string,
+        context?: Context,
+        error?: ErrorLog,
+        request?: Request,
+        response?: Response,
+        data?: Data
+    }
+    
+    export interface ChildLogRootValues {
+        module?: string,
+        category?: Enums.Category,
+        context?: {
+            piiLevel?: Enums.PiiLevelEnum
+        },
+        error?:  {
+            isHumanRelated?: boolean
+        }
+    }
+    
+    export interface ChildLogDataValues {
+        [key: string]: any;
+    }
+    
+    export interface ExpandedLog extends Log {
+        [key: string]: any;
+    }
+
+    export interface ApiContextMiddleware {
+        generateContextIfMissing?: boolean
+    }
+    
+    export interface SourceContextMiddleware extends ApiContextMiddleware {
+        context?: Context
+    }
 }
